add more trips and stages
